Add a normal hashmap implementation.
Make dashmap an optional feature.
