+++ Cargo.toml
@@ -12,4 +12,4 @@ license = "MIT"
 
 
 [dependencies]
-dashmap = "4.0.2"
+dashmap = { version = "4.0.2", optional = true }
