+++ README.md
@@ -11,8 +11,8 @@ use cachemap::CacheMap;
 	
 let m = CacheMap::new();
 
-let fst = m.cache("key", || 5u32).as_ref();
-let snd = m.cache("key", || 7u32).as_ref();
+let fst = m.cache("key", || 5u32);
+let snd = m.cache("key", || 7u32);
 
 assert_eq!(*fst, *snd);
 assert_eq!(*fst, 5u32);
@@ -22,11 +22,7 @@ assert_eq!(*fst, 5u32);
 
 - can cache values concurrently (using `&CacheMap<K,V>` rather than `&mut CacheMap<K,V>`)
 - returned references use the map's lifetime, so clients can avoid smart pointers
-- clients can optionally get `Arc<V>` pointers, in case values need to outlive the map
-- values can be addes as `Arc<V>`, allowing unsized values, and re-using `Arc<V>`s from elsewhere
-
-## MisFeatures 💧
-
-> A cache with a bad policy is another name for a memory leak
-
-this map provides only one way to remove things from the cache: drop the entire map.
+- clients can optionally enable the `dashmap` feature, which uses `dashmap`
+  internally and allows:
+  - getting `Arc<V>` pointers, in case values need to outlive the map
+  - adding `Arc<V>` directly, allowing unsized values, and re-using `Arc<V>`s from elsewhere
